+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Dude color from message
 - Dude life time
 - Dude emotes spitter
+- Message box animation, refactoring